clean up code + fluentblocks color update
