+++ packages/react-components/theme-designer/package.json
@@ -34,7 +34,7 @@
     "tslib": "^2.1.0",
     "@fluentui/react-components": "^9.0.1",
     "@fluentui/react-icons": "^2.0.172-rc.8",
-    "@fluent-blocks/colors": "9.1.0",
+    "@fluent-blocks/colors": "9.2.0",
     "codesandbox-import-utils": "2.2.3",
     "@types/dedent": "0.7.0",
     "@fluentui/react-alert": "9.0.0-beta.5"
