+++ packages/react-components/theme-designer/src/components/AccessibilityChecker/AccessibilityChecker.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { AccessibilityList } from './AccessibilityList';
-import { Vec3, hex_to_sRGB } from '@fluent-blocks/colors';
-import { contrast } from '../../utils/csswg';
+import { contrast, hex_to_sRGB, Vec3 } from '@fluent-blocks/colors';
 import { Caption1, Theme } from '@fluentui/react-components';
 
 export interface AccessibilityCheckerProps {
