+++ packages/react-components/theme-designer/src/components/ColorTokens/ColorTokens.tsx
@@ -6,7 +6,7 @@ import { Button, Caption1 } from '@fluentui/react-components';
 import { Brands, BrandVariants, teamsLightTheme } from '@fluentui/react-theme';
 import { OverridableTokenBrandColors } from './OverridableTokenBrandColors';
 import { brandTeams } from '../../utils/brandColors';
-
+import { getCurrentOverride, useColorOverrideReducer } from './useColorOverrideReducer';
 import type { DispatchTheme, ReducerState } from '../../useThemeDesignerReducer';
 
 export interface ColorTokensProps {
@@ -15,21 +15,6 @@ export interface ColorTokensProps {
   appState: ReducerState;
   dispatchState: React.Dispatch<DispatchTheme>;
 }
-
-export type ColorOverrides = {
-  teamsLight: Record<string, Brands>;
-  teamsDark: Record<string, Brands>;
-  webLight: Record<string, Brands>;
-  webDark: Record<string, Brands>;
-  customLight: Record<string, Brands>;
-  customDark: Record<string, Brands>;
-};
-
-export type DispatchColorOverrides = {
-  type: string;
-  colorToken?: string;
-  newValue?: Brands;
-};
 
 const useStyles = makeStyles({
   root: {},
@@ -47,90 +32,7 @@ export const ColorTokens: React.FunctionComponent<ColorTokensProps> = props => {
 
   const { brand, appState, dispatchState } = props;
 
-  const colorOverrideReducer: (
-    state: ColorOverrides,
-    action: { type: string; colorToken?: string; newValue?: Brands },
-  ) => ColorOverrides = (state, action) => {
-    switch (action.type) {
-      case 'Add Override':
-        if (!action.colorToken || !action.newValue) {
-          return state;
-        }
-        dispatchState({ type: 'Override', overrides: { [action.colorToken]: brand[action.newValue] } });
-        switch (appState.themeLabel) {
-          case 'Teams Light':
-            return { ...state, teamsLight: { ...state.teamsLight, [action.colorToken]: action.newValue } };
-          case 'Teams Dark':
-            return { ...state, teamsDark: { ...state.teamsDark, [action.colorToken]: action.newValue } };
-          case 'Web Light':
-            return { ...state, webLight: { ...state.webLight, [action.colorToken]: action.newValue } };
-          case 'Web Dark':
-            return { ...state, webDark: { ...state.webDark, [action.colorToken]: action.newValue } };
-          case 'Custom':
-            if (!appState.isDark) {
-              return { ...state, customLight: { ...state.customLight, [action.colorToken]: action.newValue } };
-            } else {
-              return { ...state, customDark: { ...state.customDark, [action.colorToken]: action.newValue } };
-            }
-          default:
-            return state;
-        }
-      case 'Reset Overrides':
-        dispatchState({ type: 'Override' });
-        switch (appState.themeLabel) {
-          case 'Teams Light':
-            return { ...state, teamsLight: {} };
-          case 'Teams Dark':
-            return { ...state, teamsDark: {} };
-          case 'Web Light':
-            return { ...state, webLight: {} };
-          case 'Web Dark':
-            return { ...state, webDark: {} };
-          case 'Custom':
-            if (!appState.isDark) {
-              return { ...state, customLight: {} };
-            } else {
-              return { ...state, customDark: {} };
-            }
-          default:
-            return state;
-        }
-      case 'Reset Custom Overrides':
-        return { ...state, customLight: {}, customDark: {} };
-      default:
-        return state;
-    }
-  };
-
-  const currOverride = () => {
-    switch (appState.themeLabel) {
-      case 'Teams Light':
-        return colorOverride.teamsLight;
-      case 'Teams Dark':
-        return colorOverride.teamsDark;
-      case 'Web Light':
-        return colorOverride.webLight;
-      case 'Web Dark':
-        return colorOverride.webDark;
-      case 'Custom':
-        if (!appState.isDark) {
-          return colorOverride.customLight;
-        } else {
-          return colorOverride.customDark;
-        }
-      default:
-        return colorOverride.customLight;
-    }
-  };
-
-  const [colorOverride, dispatchColorOverride] = React.useReducer(colorOverrideReducer, {
-    teamsLight: {},
-    teamsDark: {},
-    webLight: {},
-    webDark: {},
-    customLight: {},
-    customDark: {},
-  });
+  const [colorOverride, dispatchColorOverride] = useColorOverrideReducer(appState, brand, dispatchState);
 
   const handleButtonClick = () => {
     dispatchColorOverride({ type: 'Reset Overrides' });
@@ -149,7 +51,7 @@ export const ColorTokens: React.FunctionComponent<ColorTokensProps> = props => {
       <ColorTokensList
         brand={brand}
         brandColors={brandColors}
-        colorOverride={currOverride()}
+        colorOverride={getCurrentOverride(appState, colorOverride)}
         dispatchColorOverride={dispatchColorOverride}
       />
     </div>
